refactor for callbacks
